feat(story): create story file from epic [Story-7-9-6]
- The story directory already exists. Let me check the current story file and load the remaining instruction shard and required artifacts.
- The story file already exists with comprehensive content. Let me execute the workflow steps to validate and finalize it. Starting with Step 0 (load guidelines) and checking sprint status.
- The story is already in `ready-for-dev` status in sprint-status.yaml. The story file exists with comprehensive content. Let me now validate it against the workflow requirements.
- Required: `std::unique_ptr`, `std::chrono`, forward slashes, `#pragma once`
- The story already exists with comprehensive content. Let me now validate and enhance it with previous story intelligence from 7-9-5.

+++ src/source/Core/CameraMove.cpp
@@ -160,14 +160,16 @@ bool CCameraMove::LoadCameraWalkScript(const std::wstring& filename)
         return false;
     }
 
-    size_t waypointCount = 0;
-    if (fread(&waypointCount, sizeof(size_t), 1, fileHandle.get()) != 1)
+    // File format stores a 4-byte count (Win32 size_t = 4 bytes).
+    // Use uint32_t — size_t is 8 bytes on arm64, reading garbage.
+    uint32_t waypointCount = 0;
+    if (fread(&waypointCount, sizeof(uint32_t), 1, fileHandle.get()) != 1)
     {
         return false;
     }
 
     m_listWayPoint.reserve(waypointCount);
-    for (size_t index = 0; index < waypointCount; ++index)
+    for (uint32_t index = 0; index < waypointCount; ++index)
     {
         auto waypoint = std::make_unique<WAYPOINT>();
         if (fread(waypoint.get(), sizeof(WAYPOINT), 1, fileHandle.get()) != 1)
@@ -199,10 +201,10 @@ bool CCameraMove::SaveCameraWalkScript(const std::wstring& filename)
     }
 
     const DWORD signature = 0x00535743;
-    const size_t waypointCount = m_listWayPoint.size();
+    const uint32_t waypointCount = static_cast<uint32_t>(m_listWayPoint.size());
 
     if (fwrite(&signature, sizeof(DWORD), 1, fileHandle.get()) != 1 ||
-        fwrite(&waypointCount, sizeof(size_t), 1, fileHandle.get()) != 1)
+        fwrite(&waypointCount, sizeof(uint32_t), 1, fileHandle.get()) != 1)
     {
         return false;
     }
